requirements.txt
flask>=2.3.0
google-genai>=0.3.0
ollama>=0.3.0
python-dotenv
pydantic-ai>=0.0.13
tavily-python>=0.3.0
httpx>=0.25.0